refactor: CarController 코드를 각 역할에 맞게 분리 -  domain 부분은 Race.js, validate 부분은 validation/, controller 부분은 App.js으로 분리
